보드게시판 기능끝남. Excepction기능 구현중

+++ src/main/java/com/mysite/sitebackend/board/coin/service/CoinBoardService.java
@@ -1,7 +1,6 @@
 package com.mysite.sitebackend.board.coin.service;
 
 
-import antlr.StringUtils;
 import com.mysite.sitebackend.board.coin.dao.CoinBoardCommentRepository;
 import com.mysite.sitebackend.board.coin.dao.CoinBoardRepository;
 import com.mysite.sitebackend.board.coin.domain.CoinBoard;
@@ -10,18 +9,10 @@ import com.mysite.sitebackend.board.coin.dto.CoinBoardDto;
 import com.mysite.sitebackend.board.coin.dto.CoinBoardListDto;
 import com.mysite.sitebackend.board.dto.BoardInput;
 import lombok.RequiredArgsConstructor;
-import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.stereotype.Service;
 import org.modelmapper.ModelMapper;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
 
-import javax.persistence.Column;
-import javax.persistence.ManyToOne;
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.List;
 import java.util.Optional;
@@ -38,16 +29,16 @@ public class CoinBoardService {
     String formatedNow = now.format(formatter);
 
     //게시글 작성하기
-    public void boardPost(BoardInput boardInput){
-        CoinBoard c1 = new CoinBoard();
-        c1.setSubject(boardInput.getSubject());
-        c1.setContents(boardInput.getContents());
-        c1.setAuthor(boardInput.getAuthor());
-        c1.setViews(0);
-        c1.setDate(formatedNow);
-        this.boardRepository.save(c1);
+    public boolean boardPost(BoardInput boardInput) throws Exception{
+            CoinBoard c1 = new CoinBoard();
+            c1.setSubject(boardInput.getSubject());
+            c1.setContents(boardInput.getContents());
+            c1.setAuthor(boardInput.getAuthor());
+            c1.setViews(0);
+            c1.setDate(formatedNow);
+            this.boardRepository.save(c1);
+            return true;
     }
-
     //댓글 작성하기
     public boolean commentPost(BoardInput boardInput){
         Optional<CoinBoard> optionalCoinBoard = this.boardRepository.findById(boardInput.getId());
@@ -73,7 +64,6 @@ public class CoinBoardService {
                         .collect(Collectors.toList());
         return coinBoardListDto;
     }
-
     //게시글 1개불러오기
     public CoinBoardDto findByIdToBoard(BoardInput boardInput){
         CoinBoard board = boardRepository.findById(boardInput.getId()).get();
@@ -81,42 +71,85 @@ public class CoinBoardService {
         CoinBoardDto CoinboardDto = modelMapper.map(board, CoinBoardDto.class);
         return CoinboardDto;
     }
-
     //게시글 1개의 댓글 전체목록 불러오기
     public List<CoinBoardComment> findByIdToComment(BoardInput boardInput){
         Optional<CoinBoard> optionalCoinBoard = this.boardRepository.findById(boardInput.getId());
         optionalCoinBoard.isPresent() ;
             List<CoinBoardComment> coinBoardComments = commentRepository.findAllByBoardIndex(boardInput.getId());
-
         return coinBoardComments;
     }
 
     //게시글 수정
-    public CoinBoardDto findByIdToPatch(BoardInput boardInput){
+    public boolean boardPatch(BoardInput boardInput){
         Optional<CoinBoard> opBoard = boardRepository.findById(boardInput.getId());
         if(opBoard.isPresent()){
-            CoinBoard board = opBoard.get();
-            board.setSubject(boardInput.getSubject());
-            board.setContents(boardInput.getContents());
-            board.setAuthor(boardInput.getAuthor());
-            board.setDate(formatedNow);
-            boardRepository.save(board);
+            if(boardInput.getAuthor() == opBoard.get().getAuthor()){
+                CoinBoard board = opBoard.get();
+                board.setSubject(boardInput.getSubject());
+                board.setContents(boardInput.getContents());
+                board.setAuthor(boardInput.getAuthor());
+                board.setDate(formatedNow);
+                boardRepository.save(board);
+                return true;
+            }
+            else {
+                return false;
+            }
         }
-        CoinBoard board = boardRepository.findById(boardInput.getId()).get();
-        CoinBoardDto coinboardDto = modelMapper.map(board, CoinBoardDto.class);
-        return coinboardDto;
+        else{
+            return false;
+        }
+    }
+    //댓글 수정
+    public boolean commentPatch(BoardInput boardInput){
+        Optional<CoinBoardComment> opComment = commentRepository.findById(boardInput.getId());
+        if(opComment.isPresent()){
+            if(boardInput.getAuthor() == opComment.get().getAuthor()){
+                CoinBoardComment comment = opComment.get();
+                comment.setContents(boardInput.getContents());
+                comment.setAuthor(boardInput.getAuthor());
+                comment.setDate(formatedNow);
+                comment.setBoardIndex(comment.getBoardIndex());
+                commentRepository.save(comment);
+                return true;
+            }
+            else{
+                return false;
+            }
+        }
+        else{
+            return false;
+        }
+
     }
 
     //게시글 삭제
-    public String boardDelete(BoardInput boardInput){
-        this.boardRepository.deleteById(boardInput.getId());
-        this.commentRepository.deleteAllByBoardIndex(boardInput.getId());
-        return boardInput.getId() + "번 게시판의 삭제가 완료되었습니다.";
+    public boolean boardDelete(BoardInput boardInput){
+        Optional<CoinBoard> optionalCoinBoard = this.boardRepository.findById(boardInput.getId());
+        if(optionalCoinBoard.isPresent()){
+            if(boardInput.getAuthor() == optionalCoinBoard.get().getAuthor()){
+                this.boardRepository.deleteById(boardInput.getId());
+                this.commentRepository.deleteAllByBoardIndex(boardInput.getId());
+                return true;
+            }
+            else {
+                return false;
+            }
+        }
+        else{
+            return false;
+        }
     }
     // 댓글삭제
-    public String commnetDelete(BoardInput boardInput){
-        this.commentRepository.deleteById(boardInput.getId());
-        return boardInput.getId() + "번 댓글의 삭제가 완료되었습니다.";
+    public boolean commnetDelete(BoardInput boardInput){
+        Optional<CoinBoardComment> optionalCoinBoardComment = this.commentRepository.findById(boardInput.getId());
+        if(optionalCoinBoardComment.isPresent()){
+            this.commentRepository.deleteById(boardInput.getId());
+            return true;
+        }
+        else{
+            return false;
+        }
     }
 
 }
